Tests: async into own module
@pytest.mark.django_db(transaction=True) was applying to whole module. Which it shouldn't have been. And seemingly couldn't be told otherwise.

Importantly:
archiver/tests.py ...............
composer/tests.py ............
folk_rnn_site/tests.py .....
functional_tests/tests.py .
33 passed in 3.53 seconds
