+++ folk_rnn_site/folk_rnn_site/tests.py
@@ -4,6 +4,21 @@ from django.contrib.staticfiles import finders
 
 from folk_rnn_site.models import ABCModel
 
+# Input, output as per https://github.com/tobyspark/folk-rnn/commit/381184a2d6659a47520cedd6d4dfa7bb1c5189f7
+FOLKRNN_IN = {'rnn_model_name': 'with_repeats.pickle', 'seed': 42, 'temp': 1, 'meter': '', 'key': '', 'start_abc': ''}
+FOLKRNN_OUT_RAW = '''M:4/4 K:Cdor c 3 d c 2 B G | G F F 2 G B c d | c 3 d c B G A | B G G F G 2 f e | d 2 c d c B G A | B G F G B F G F | G B c d c d e f | g b f d e 2 e f | d B B 2 B 2 d c | B G G 2 B G F B | d B B 2 c d e f | g 2 f d g f d c | d g g 2 f d B c | d B B 2 B G B c | d f d c B 2 d B | c B B G F B G B | d 2 c d d 2 f d | g e c d e 2 f g | f d d B c 2 d B | d c c B G B B c | d 2 c d d 2 f d | c d c B c 2 d f | g b b 2 g a b d' | c' d' b g f d d f |'''
+FOLKRNN_OUT = '''X:1
+T:Folk RNN Candidate Tune No1
+M:4/4
+K:Cdor
+c3d c2BG|GFF2 GBcd|c3d cBGA|BGGF G2fe|
+d2cd cBGA|BGFG BFGF|GBcd cdef|gbfd e2ef|
+dBB2 B2dc|BGG2 BGFB|dBB2 cdef|g2fd gfdc|
+dgg2 fdBc|dBB2 BGBc|dfdc B2dB|cBBG FBGB|
+d2cd d2fd|gecd e2fg|fddB c2dB|dccB GBBc|
+d2cd d2fd|cdcB c2df|gbb2 gabd'|c'd'bg fddf|
+'''
+
 ABC_TITLE = 'Test Tune'
 ABC_BODY = 'A B C'
