+++ folk_rnn_site/pytest.ini
@@ -1,3 +1,3 @@
 [pytest]
 DJANGO_SETTINGS_MODULE = folk_rnn_site.settings
-python_files = tests.py
+python_files = tests.py test_*.py *_tests.py
